refactor(files.service): move all files services to folder
