+++ src/services/files/index.ts
@@ -1,0 +1,5 @@
+export * from './files.service.js';
+export * from './files.worker.service.js';
+export * from './linux-files.service.js';
+export * from './mac-files.service.js';
+export * from './windows-files.service.js';
