+++ src/services/index.ts
@@ -1,10 +1,8 @@
 export * from './console.service.js';
-export * from './files.service.js';
 export * from './https.service.js';
-export * from './linux-files.service.js';
-export * from './mac-files.service.js';
 export * from './results.service.js';
 export * from './spinner.service.js';
 export * from './stream.service.js';
 export * from './update.service.js';
-export * from './windows-files.service.js';
+
+export * from './files/index.js';
